requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==2.1.15
django-allauth==0.42.0
django-allauth-bootstrap==0.2
django-appconf==1.0.4
django-bootstrap-form==3.4
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.0.1
django-imagekit==4.0.2
django-random-queryset==0.1.3
idna==2.9
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.2
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
